Initial transformation of Redis test cases
